Rename `OnPrem` -> `SelfHosted`
