+++ Deepgram/Clients/Interfaces/v1/ISelfHostedClient.cs
@@ -1,0 +1,50 @@
+﻿// Copyright 2021-2024 Deepgram .NET SDK contributors. All Rights Reserved.
+// Use of this source code is governed by a MIT license that can be found in the LICENSE file.
+// SPDX-License-Identifier: MIT
+
+using Deepgram.Models.SelfHosted.v1;
+
+namespace Deepgram.Clients.Interfaces.v1;
+
+/// <summary>
+/// Implements version 1 of the OnPrem Client.
+/// </summary>
+/// <param name="apiKey">Required DeepgramApiKey</param>
+/// <param name="deepgramClientOptions"><see cref="DeepgramHttpClientOptions"/> for HttpClient Configuration</param>
+public interface ISelfHostedClient
+{
+    /// <summary>
+    /// get a list of credentials associated with project
+    /// </summary>
+    /// <param name="projectId">Id of project</param>
+    /// <returns><see cref="CredentialsResponse"/></returns>
+    public Task<CredentialsResponse> ListCredentials(string projectId, CancellationTokenSource? cancellationToken = default,
+        Dictionary<string, string>? addons = null, Dictionary<string, string>? headers = null);
+
+    /// <summary>
+    /// Get credentials for the project that is associated with credential ID
+    /// </summary>
+    /// <param name="projectId">Id of project</param>
+    /// <param name="credentialsId">Id of credentials</param>
+    /// <returns><see cref="CredentialResponse"/></returns>
+    public Task<CredentialResponse> GetCredentials(string projectId, string credentialsId, CancellationTokenSource? cancellationToken = default,
+        Dictionary<string, string>? addons = null, Dictionary<string, string>? headers = null);
+
+    /// <summary>
+    /// Remove credentials in the project associated with the credentials ID
+    /// </summary>
+    /// <param name="projectId">Id of project</param>
+    /// <param name="credentialsId">Id of credentials</param>
+    /// <returns><see cref="MessageResponse"/></returns>
+    public Task<MessageResponse> DeleteCredentials(string projectId, string credentialsId, CancellationTokenSource? cancellationToken = default,
+        Dictionary<string, string>? addons = null, Dictionary<string, string>? headers = null);
+
+    /// <summary>
+    /// Create credentials for the associated projects
+    /// </summary>
+    /// <param name="projectId">Id of project</param>
+    /// <param name="createOnPremCredentialsSchema"><see cref="CredentialsSchema"/> for credentials to be created</param>
+    /// <returns><see cref="CredentialResponse"/></returns>
+    public Task<CredentialResponse> CreateCredentials(string projectId, CredentialsSchema credentialsSchema,
+        CancellationTokenSource? cancellationToken = default, Dictionary<string, string>? addons = null, Dictionary<string, string>? headers = null);
+}
