+++ Deepgram/SelfHostedClient.cs
@@ -1,0 +1,19 @@
+﻿// Copyright 2021-2024 Deepgram .NET SDK contributors. All Rights Reserved.
+// Use of this source code is governed by a MIT license that can be found in the LICENSE file.
+// SPDX-License-Identifier: MIT
+
+using Deepgram.Clients.SelfHosted.v1;
+using Deepgram.Models.Authenticate.v1;
+
+namespace Deepgram;
+
+/// <summary>
+/// Implements the latest supported version of the OnPrem Client.
+/// </summary>
+public class SelfHostedClient : Client
+{
+    public SelfHostedClient(string apiKey = "", DeepgramHttpClientOptions? deepgramClientOptions = null,
+        string? httpId = null) : base(apiKey, deepgramClientOptions, httpId)
+    {
+    }
+}
